% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{construct.file.path}
\alias{construct.file.path}
\title{Construct File Path}
\usage{
construct.file.path(
  filename = NULL,
  suffix = NULL,
  extension = NULL,
  manual_file_name = NULL,
  manual_directory = NULL,
  verbose = TRUE
)
}
\arguments{
\item{filename}{The base file name to process. Default: NULL.}

\item{suffix}{The file name suffix to be appended. Default: NULL.}

\item{extension}{The file extension to be appended. Default: NULL.}

\item{manual_file_name}{An optional manual specification for the file name. Default: NULL.}

\item{manual_directory}{An optional manual specification for the directory. Default: NULL.}

\item{verbose}{Print path? Default: TRUE.}
}
\value{
A string representing the constructed file path.
}
\description{
Constructs a complete file path using either provided manual file name and directory
or defaults to processing a given filename and using the current working directory.
}
\examples{
construct.file.path(
  filename = "report", manual_file_name = NULL, manual_directory = NULL,
  extension = "txt"
)
}
