% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.Functions.R
\name{convert.tsv.data}
\alias{convert.tsv.data}
\title{convert.tsv.data}
\usage{
convert.tsv.data(df_by_read.simple.tsv, digitz = 2, na_rep = 0)
}
\arguments{
\item{df_by_read.simple.tsv}{Data frame (e.g., by \code{read.simple.tsv}).}

\item{digitz}{Number of digits when rounding up. Default: 2.}

\item{na_rep}{Replace NA? Default: 0.}
}
\description{
Fix NA issues in data frames imported by the new read.simple.tsv.
Set \code{na_rep} to NA if you want to keep NAs.
}
\seealso{
\code{\link[gtools]{na.replace}}
}
