% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{FirstCol2RowNames.as.df}
\alias{FirstCol2RowNames.as.df}
\title{FirstCol2RowNames.as.df}
\usage{
FirstCol2RowNames.as.df(Tibble, rownamecol = 1, make_names = FALSE)
}
\arguments{
\item{Tibble}{A data frame without row names (tibble style).}

\item{rownamecol}{Row name column. Default: 1.}

\item{make_names}{Call \code{make.names} to remove unusual characters. Default: FALSE.}
}
\description{
Set first column to row names.
}
