% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{qs.2.table}
\alias{qs.2.table}
\title{Convert and save a .qs file to different formats}
\usage{
qs.2.table(path, out_file = c("tsv", "csv", "csv2", "excel")[1])
}
\arguments{
\item{path}{A character string specifying the path to the \code{.qs} file. Default: none.}

\item{out_file}{A character string specifying the output file format. One of \code{"tsv"} (default),
\code{"csv"}, \code{"csv2"} (semicolon-separated), or \code{"excel"}. Default: \code{"tsv"}.}
}
\value{
The function does not return a value but writes the file to disk in the specified format.
}
\description{
This function reads in a \code{.qs} file and resaves it as either a \code{.tsv}, \code{.csv}, semicolon-separated
\code{.csv} (csv2), or Excel file based on the \code{out_file} argument.
}
