% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple.csv}
\alias{read.simple.csv}
\title{read.simple.csv}
\usage{
read.simple.csv(
  ...,
  colnames = TRUE,
  coltypes = NULL,
  wRownames = TRUE,
  NaReplace = TRUE,
  asTibble = FALSE,
  nmax = Inf
)
}
\arguments{
\item{...}{Multiple simple variables to parse.}

\item{colnames}{Are there column names? Default: TRUE.}

\item{coltypes}{What type of variables are in columns? Auto-guessing can be very slow. Default: NULL.}

\item{wRownames}{With row names? Default: TRUE.}

\item{NaReplace}{Replace NA values? Default: TRUE.}

\item{asTibble}{Load as tibble or data frame? Default: FALSE (load as data frame).}

\item{nmax}{Max number of rows to read. Default: Inf.}
}
\description{
Read in a file with Excel-style data: row names in column 1,
headers shifted. The header should start with a TAB; the first column name
should be empty.
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple.csv("path/to/my.file")
}
}
}
\seealso{
\code{\link[readr]{read_delim}}
\code{\link[gtools]{na.replace}}
}
