% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple}
\alias{read.simple}
\title{read.simple}
\usage{
read.simple(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Essentially \code{read.table()} with file/path parsing.
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple("path/to/a/file")
}
}
}
