% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple.ssv}
\alias{read.simple.ssv}
\title{read.simple.ssv}
\usage{
read.simple.ssv(
  ...,
  sep_ = " ",
  colnames = TRUE,
  wRownames = TRUE,
  NaReplace = TRUE,
  coltypes = NULL,
  asTibble = FALSE
)
}
\arguments{
\item{...}{Multiple simple variables to parse.}

\item{sep_}{Separator character. Default: ' '.}

\item{colnames}{Are there column names? Default: TRUE.}

\item{wRownames}{With row names? Default: TRUE.}

\item{NaReplace}{Replace NA values? Default: TRUE.}

\item{coltypes}{What type of variables are in columns? Auto-guessing can be very slow. Default: NULL.}

\item{asTibble}{Load as tibble or data frame? Default: FALSE (load as data frame).}
}
\description{
Space separated values. Read in a file with Excel-style data:
row names in column 1, headers shifted. The header should start with a
TAB; the first column name should be empty.
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple.ssv("path/to/my.file")
}
}
}
\seealso{
\code{\link[readr]{read_delim}}
\code{\link[gtools]{na.replace}}
}
