% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple.vec}
\alias{read.simple.vec}
\title{read.simple.vec}
\usage{
read.simple.vec(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Read each line of a file to an element of a vector (read in newline-separated values, no header!).
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple.vec("path/to/a/single.column.file")
}
}
}
