% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.Functions.R
\name{read.simple.xls}
\alias{read.simple.xls}
\title{read.simple.xls}
\usage{
read.simple.xls(
  pfn = Stringendo::kollapse(...),
  row_namePos = NULL,
  ...,
  header_ = TRUE,
  WhichSheets
)
}
\arguments{
\item{pfn}{Path and file name. Default: kollapse(...).}

\item{row_namePos}{Where is the row name? Default: NULL.}

\item{...}{Multiple simple variables to parse.}

\item{header_}{Is there a header? Default: TRUE.}

\item{WhichSheets}{Which sheets to read.}
}
\description{
Read multi-sheet Excel files. \code{row_namePos = NULL} for automatic
names. See http://readxl.tidyverse.org/.
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple.xls("path/to/my.file")
}
}
}
\seealso{
\code{\link[gdata]{read.xls}}
}
