% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple_char_list}
\alias{read.simple_char_list}
\title{read.simple_char_list}
\usage{
read.simple_char_list(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Read in a file.
}
\examples{
\dontrun{
if(interactive()){
 # read.simple_char_list("path/to/my.file")
 }
}
}
