% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{write.simple}
\alias{write.simple}
\title{Write Simple}
\usage{
write.simple(
  input_df,
  filename = substitute(input_df),
  suffix = NULL,
  extension = "tsv",
  manual_file_name = NULL,
  manual_directory = NULL,
  o = FALSE,
  v = TRUE
)
}
\arguments{
\item{input_df}{Data frame to write out. Default: None, must be provided.}

\item{filename}{The base name for the output file. Default: Name of the input data frame.}

\item{suffix}{An optional suffix to add to the filename. Default: NULL.}

\item{extension}{File extension to use. Default: 'tsv'.}

\item{manual_file_name}{Manually defined filename, overrides automatic naming. Default: NULL.}

\item{manual_directory}{Directory to save the file in, overrides default directory. Default: NULL.}

\item{o}{If TRUE, opens the file after writing on OS X using 'system(open ...)'. Default: FALSE.}

\item{v}{Print path if verbose? Default: TRUE.}
}
\value{
Outputs a .tsv file and optionally prints the length of the input data frame.
}
\description{
Writes a matrix-like R object (e.g., a data frame) to a file as tab-separated
values (.tsv). The output filename can be auto-generated from the variable's name or manually
specified. The file is saved in the specified output directory or the current working directory.
The path and variable name can be passed separately and will be concatenated to form the filename.
}
\examples{
\dontrun{
if (interactive()) {
  write.simple(input_df = myDataFrame)
}
}
}
