% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{write.simplest}
\alias{write.simplest}
\title{Append or write a vector to a standard file, one element per line.}
\usage{
write.simplest(
  vec = LETTERS[1:11],
  append = TRUE,
  header = NULL,
  prefix = Stringendo::kppws(substitute(vec), Stringendo::idate()),
  file_path = get0("path_write_simplest", ifnotfound = "./__clipboard.txt")
)
}
\arguments{
\item{vec}{A vector to be written to the file. Default: \code{LETTERS[1:11]}.}

\item{append}{Logical flag indicating whether to append the vector to the file. Default: \code{TRUE}.}

\item{header}{A string to be added to the header line (before the vector). Default: \code{NULL}.}

\item{prefix}{A prefix to the header. Default: \code{kppws(substitute(vec), idate())}.}

\item{file_path}{A string specifying the file path where the vector will be written. Default:
path stored in \code{path_write_simplest} global variable, otherwise \code{"./__clipboard.txt"}.}
}
\value{
A message indicating the length of the vector and the file path to which it was written.
}
\description{
Alternative to the clipboard. This function takes a vector and appends it
to a specified file.
}
\examples{
\dontrun{
write.simplest(letters[1:5], append = TRUE)
}

}
