% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{convert.tsv.data}
\alias{convert.tsv.data}
\title{convert.tsv.data}
\usage{
convert.tsv.data(df_by_read.simple.tsv, digitz = 2, na_rep = 0)
}
\arguments{
\item{df_by_read.simple.tsv}{Data frame (e.g. by read.simple.tsv).}

\item{digitz}{Number of digits when rounding up, Default: 2}

\item{na_rep}{Replace NA?, Default: 0}
}
\description{
Fix NA issue in dataframes imported by the new read.simple.tsv.
Set na_rep to NA if you want to keep NA-s
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[gtools]{na.replace}}
}
