% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{FirstCol2RowNames.as.df}
\alias{FirstCol2RowNames.as.df}
\title{FirstCol2RowNames.as.df}
\usage{
FirstCol2RowNames.as.df(Tibble, rownamecol = 1, make_names = FALSE)
}
\arguments{
\item{Tibble}{A dataframe without rownames (tibble style)}

\item{rownamecol}{rowname column, Default: 1}

\item{make_names}{call make.names to remove weird characters, Default: FALSE}
}
\description{
Set First Col to Row Names
}
