% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{FirstCol2RowNames}
\alias{FirstCol2RowNames}
\title{FirstCol2RowNames}
\usage{
FirstCol2RowNames(Tibble, rownamecol = 1, make_names = FALSE, as.df = TRUE)
}
\arguments{
\item{Tibble}{A dataframe without rownames (tibble style)}

\item{rownamecol}{rowname column, Default: 1}

\item{make_names}{call make.names to remove weird characters, Default: FALSE}

\item{as.df}{Convert tibble to data frame? Default: TRUE}
}
\description{
Set First Col to Row Names
}
