% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{FirstCol2RowNames}
\alias{FirstCol2RowNames}
\title{FirstCol2RowNames}
\usage{
FirstCol2RowNames(Tibble, rownamecol = 1, make_names = FALSE, as.df = TRUE)
}
\arguments{
\item{Tibble}{A data frame without row names (tibble style).}

\item{rownamecol}{Row name column. Default: 1.}

\item{make_names}{Call \code{make.names} to remove unusual characters. Default: FALSE.}

\item{as.df}{Convert tibble to data frame? Default: TRUE.}
}
\description{
Set first column to row names.
}
