% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple.csv.named.vector}
\alias{read.simple.csv.named.vector}
\title{read.simple.csv.named.vector}
\usage{
read.simple.csv.named.vector(
  file,
  sep = ";",
  col_names = FALSE,
  value_col = 2,
  name_col = 1,
  ...
)
}
\arguments{
\item{file}{Path to the *.csv file.}

\item{sep}{Separator character, Default: ';' alternative: ','.}

\item{col_names}{Are there column names?, Default: TRUE}

\item{value_col}{Column number of the values in the input data frame. Default: 2}

\item{name_col}{Column number of the names in the input data frame. Default: 1}

\item{...}{Additional arguments passed to \code{\link[readr]{read_csv}} or read_csv2.}
}
\description{
Read in a data frame (csv), and extract a value and a name column, and convert them
to a named vector. By default, it assumes the names in the first column and the values
excel style named vectors, names in col1,
headers SHIFTED. The header should start with a TAB / First column name
should be empty.
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple.csv.named.vector("path/to/my.file.csv")
}
}
}
\seealso{
\code{\link[readr]{read_delim}}
}
