% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple.table}
\alias{read.simple.table}
\title{read.simple.table}
\usage{
read.simple.table(..., colnames = TRUE, coltypes = NULL)
}
\arguments{
\item{...}{Multiple simple variables to parse.}

\item{colnames}{Are there column names? Default: TRUE.}

\item{coltypes}{What type of variables are in columns? Auto-guessing can be very slow. Default: NULL.}
}
\description{
Read a file. Default: header defines column names, no row names.
For row names give the column number with row names, e.g., 1. The header should start
with a TAB; the first column name should be empty.
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple.table("path/to/my.file")
}
}
}
\seealso{
\code{\link[readr]{read_delim}}
\code{\link[gtools]{na.replace}}
}
