% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple.tsv.named.vector}
\alias{read.simple.tsv.named.vector}
\title{read.simple.tsv.named.vector}
\usage{
read.simple.tsv.named.vector(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Read in a file with Excel-style named vectors, names in column 1,
headers shifted. The header should start with a TAB; the first column name
should be empty.
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple.tsv.named.vector("path/to/my.file")
}
}
}
\seealso{
\code{\link[readr]{read_delim}}
}
