% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple.tsv}
\alias{read.simple.tsv}
\title{read.simple.tsv}
\usage{
read.simple.tsv(
  ...,
  sep_ = "\\t",
  colnames = TRUE,
  wRownames = TRUE,
  coltypes = NULL,
  NaReplace = TRUE,
  asTibble = FALSE
)
}
\arguments{
\item{...}{Multiple simple variables to parse.}

\item{sep_}{Separator character. Default: '\\t'.}

\item{colnames}{Are there column names? Default: TRUE.}

\item{wRownames}{With row names? Default: TRUE.}

\item{coltypes}{What type of variables are in columns? Auto-guessing can be very slow. Default: NULL.}

\item{NaReplace}{Replace NA values? Default: TRUE.}

\item{asTibble}{Load as tibble or data frame? Default: FALSE (load as data frame).}
}
\description{
Read in a file with Excel-style data: row names in column 1,
headers shifted. The header should start with a TAB; the first column name
should be empty.
}
\examples{
\dontrun{
if (interactive()) {
  # read.simple.tsv("path/to/my.file")
}
}
}
\seealso{
\code{\link[readr]{read_delim}}
\code{\link[gtools]{na.replace}}
}
