% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{read.simple.xls}
\alias{read.simple.xls}
\title{read.simple.xls}
\usage{
read.simple.xls(
  pfn = kollapse(...),
  row_namePos = NULL,
  ...,
  header_ = TRUE,
  WhichSheets
)
}
\arguments{
\item{pfn}{Path and File name, Default: kollapse(...)}

\item{row_namePos}{Where is the rowname, Default: NULL}

\item{...}{Multiple simple variables to parse.}

\item{header_}{Is there header? Default: TRUE}

\item{WhichSheets}{Which sheets to read in}
}
\description{
Read multi-sheet excel files. row_namePos = NULL for automatic
names Look into: http://readxl.tidyverse.org/.
}
\examples{
\dontrun{
if(interactive()){
 # read.simple.xls("path/to/my.file")
 }
}
}
\seealso{
\code{\link[gdata]{read.xls}}
}
