% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{write.simple.append}
\alias{write.simple.append}
\title{write.simple.append}
\usage{
write.simple.append(
  input_df,
  extension = "tsv",
  filename = substitute(input_df),
  suffix = NULL,
  ManualName = "",
  o = FALSE,
  ...
)
}
\arguments{
\item{input_df}{Data frame to write out.}

\item{extension}{File extension to add, Default: 'tsv'}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{ManualName}{A string for a manually defined filename. Default: ''}

\item{o}{Set to TRUE to open file after writing out using 'system(open ...)' on OS X., Default: FALSE}

\item{...}{Multiple simple variables to parse.}
}
\description{
Append an R-object WITHOUT ROWNAMES, to an existing .tsv file of
the same number of columns. Your output filename will be either the
variable's name. The output file will be located in "OutDir" specified by
you at the beginning of the script, or under your current working directory.
You can pass the PATH and VARIABLE separately (in order), they will be
concatenated to the filename.
}
\examples{
\dontrun{
if(interactive()){
 # write.simple.append(my.data.frame)
 }
}
}
