% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{write.simple.vec}
\alias{write.simple.vec}
\title{write.simple.vec}
\usage{
write.simple.vec(input_vec, extension = "vec", ManualName = "", o = FALSE, ...)
}
\arguments{
\item{input_vec}{Vector to write out.}

\item{extension}{File extension to add, Default: 'vec'}

\item{ManualName}{A string for a manually defined filename. Default: ''}

\item{o}{Set to TRUE to open file after writing out using 'system(open ...)' on OS X., Default: FALSE}

\item{...}{Multiple simple variables to parse.}
}
\description{
Write out a vector-like R-object to a file with as newline
separated values (.vec). Your output filename will be either the variable's
name. The output file will be located in "OutDir" specified by you at the
beginning of the script, or under your current working directory. You can
pass the PATH and VARIABLE separately (in order), they will be concatenated
to the filename.
}
\examples{
\dontrun{
if(interactive()){
 # write.simple.vec(my.vector)
 }
}
}
