% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{write.simple.vec}
\alias{write.simple.vec}
\title{Write Simple Vector}
\usage{
write.simple.vec(
  input_vec,
  filename = substitute(input_vec),
  suffix = NULL,
  extension = "vec",
  manual_file_name = NULL,
  manual_directory = NULL,
  o = FALSE,
  v = TRUE
)
}
\arguments{
\item{input_vec}{Vector to write out. Default: None, must be provided.}

\item{filename}{The base name for the output file. Default: Name of the input vector.}

\item{suffix}{An optional suffix to add to the filename. Default: NULL.}

\item{extension}{File extension to use. Default: 'vec'.}

\item{manual_file_name}{Manually defined filename, overrides automatic naming. Default: NULL.}

\item{manual_directory}{Directory to save the file in, overrides default directory. Default: NULL.}

\item{o}{If TRUE, opens the file after writing on OS X using 'system(open ...)'. Default: FALSE.}

\item{v}{Print path if verbose? Default: TRUE.}
}
\value{
Outputs a .vec file and optionally prints the length of the input vector.
}
\description{
Writes a vector-like R object to a file as newline-separated values (.vec).
The output filename can be auto-generated from the variable's name or manually specified. The file
is saved in the specified output directory or the current working directory. The path and variable
name can be passed separately and will be concatenated to form the filename.
}
\examples{
\dontrun{
if (interactive()) {
  write.simple.vec(input_vec = myVector)
}
}
}
