% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{write.simple.xlsx}
\alias{write.simple.xlsx}
\title{Write Simple XLSX}
\usage{
write.simple.xlsx(
  named_list,
  rowname_column = 1,
  filename = substitute(named_list),
  suffix = NULL,
  manual_file_name = NULL,
  manual_directory = NULL,
  o = FALSE,
  gzip = FALSE,
  TabColor = "darkgoldenrod1",
  HeaderLineColor = "darkolivegreen3",
  HeaderCex = 12,
  Creator = "",
  HeaderCharStyle = c("bold", "italic", "underline")[1],
  has_row_names = TRUE,
  FreezeFirstRow = TRUE,
  FreezeFirstCol = FALSE,
  v = TRUE
)
}
\arguments{
\item{named_list}{A list of data frames or matrices to write out.
Default: No default value; a list must be provided.}

\item{rowname_column}{The column name or index to use as row names in the Excel file.
Default: 1.}

\item{filename}{The base name for the output file, derived from the 'named_list' variable if not specified.
Default: Derived using \code{substitute(named_list)}.}

\item{suffix}{A suffix to be added to the output filename. Default: NULL.}

\item{manual_file_name}{Manually defined filename, overrides automatic naming. Default: NULL.}

\item{manual_directory}{Directory to save the file in, overrides default directory. Default: NULL.}

\item{o}{Logical; if TRUE, opens the file after writing using the system's default application.
Default: FALSE.}

\item{gzip}{Compress the file after saving? Default: FALSE.}

\item{TabColor}{Color for the tabs in Excel. Default: 'darkgoldenrod1'.}

\item{HeaderLineColor}{Color for the header line. Default: 'darkolivegreen3'.}

\item{HeaderCex}{Font size for the header. Default: 12.}

\item{Creator}{The creator of the Excel document. Default: ''.}

\item{HeaderCharStyle}{Character style for the header (e.g., 'bold', 'italic', 'underline').
Default: 'bold'.}

\item{has_row_names}{Logical; if set to FALSE, converts the first column to row names. Default: TRUE.}

\item{FreezeFirstRow}{Logical; if TRUE, freezes the first row in Excel. Default: TRUE.}

\item{FreezeFirstCol}{Logical; if TRUE, freezes the first column in Excel. Default: FALSE.}

\item{v}{Print path if verbose? Default: TRUE.}
}
\description{
Write out a list of matrices or data frames with row and column names
to an Excel (.xlsx) file. The output filename is generated based on the provided parameters
and stored in the specified output directory or the current working directory.
The function offers various styling and formatting options for the Excel file.
}
\examples{
\dontrun{
if (interactive()) {
  # Example usage:
  # write.simple.xlsx(my.list.of.data.frames, rowname_column = "gene")
}
}
}
\seealso{
\code{\link[openxlsx]{write.xlsx}}
}
