% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{write.simple.xlsx}
\alias{write.simple.xlsx}
\title{write.simple.xlsx}
\usage{
write.simple.xlsx(
  named_list,
  filename = substitute(named_list),
  suffix = NULL,
  o = FALSE,
  TabColor = "darkgoldenrod1",
  HeaderLineColor = "darkolivegreen3",
  HeaderCex = 12,
  Creator = "",
  HeaderCharStyle = c("bold", "italic", "underline")[1],
  row_names = TRUE,
  ...
)
}
\arguments{
\item{named_list}{A list of data frames to write out}

\item{suffix}{A suffix added to the filename, Default: NULL}

\item{o}{Set to TRUE to open file after writing out using 'system(open ...)' on OS X., Default: FALSE}

\item{TabColor}{Tab Color in Excel, Default: 'darkgoldenrod1'}

\item{HeaderLineColor}{Header line color, Default: 'darkolivegreen3'}

\item{HeaderCex}{Header color, Default: 12}

\item{Creator}{Creator, Default: ''}

\item{HeaderCharStyle}{Header character style, Default: c("bold", "italic", "underline")\link{1}}

\item{row_names}{Have rownames? Default: TRUE}

\item{...}{Multiple simple variables to parse.}

\item{fname}{A string for a manually defined filename. Default: substitute(named_list)}
}
\description{
Write out a list of matrices/ data frames WITH ROW- AND COLUMN-
NAMES to a file with as an Excel (.xslx) file. Your output filename will be
either the variable's name. The output file will be located in "OutDir"
specified by you at the beginning of the script, or under your current
working directory. You can pass the PATH and VARIABLE separately (in
order), they will be concatenated to the filename.
}
\examples{
\dontrun{
if(interactive()){
 # write.simple.xlsx(my.list.of.data.frames)
 }
}
}
\seealso{
\code{\link[openxlsx]{write.xlsx}}
}
