% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriter.R
\name{write.simplest}
\alias{write.simplest}
\title{Append or write a vector to standard file, one element per line.}
\usage{
write.simplest(
  vec = LETTERS[1:11],
  append = TRUE,
  file_path = "/groups/knoblich/Projects/connectomics/Analysis/__clipboard.txt"
)
}
\arguments{
\item{vec}{A vector to be written to the file. Default: \code{LETTERS[1:11]}.}

\item{append}{Logical flag indicating whether to append the vector to the file. Default: \code{TRUE}.}

\item{file_path}{A string specifying the file path where the vector will be written. Default:
\code{"/groups/knoblich/Projects/connectomics/Analysis/__clipboard.txt"}.}
}
\value{
A message indicating the length of the vector and the file path to which it was written.
}
\description{
Alternative to clipboard. This function takes a vector and appends it
to a specified file.
}
\examples{
\dontrun{
 write.simplest(letters[1:5], append = TRUE)
 }

}
